Combine blackbox and marginclient
